Move crates/rpc/rpc-eth-api/src/helpers/types.rs up one dir
